Separate controllers from routes
