+++ practice-app/backend/routes/trading-equipments.js
@@ -1,7 +1,6 @@
 const express = require('express')
-const secret = require('../secrets')
-const request= require('request');
 const router = express.Router()
+const teControllers = require('../controllers/trading-equipments')
 
 /* 
  * Retrieves a certain exchange rate. Takes two 
@@ -9,54 +8,12 @@ const router = express.Router()
  * uses alphavantage api
  * currency parameters should be entered by their fx codes
  */ 
-router.get("/:from-:to", (req, res) => {
-    let params = req.params
-    request(`https://www.alphavantage.co/query?function=CURRENCY_EXCHANGE_RATE&from_currency=${params.from}&to_currency=${params.to}&apikey=${secret.alphaKey.key}`, 
-    { json: true }, (err, resp, body) => {
-  
-        if (err) { // if external api returns error
-            res.send({
-                Error: "System is under maintanence."
-            });
-        }
-        else if (body.hasOwnProperty('Realtime Currency Exchange Rate')){//when everything is normal sends a rate and its value as JSON
-            res.send({
-                rate: `${params['from']}/${params['to']}`,
-                value: body['Realtime Currency Exchange Rate']['5. Exchange Rate']
-            });
-        }
-        else if (body.hasOwnProperty('Error Message')){//when user gives wrong parameters
-            res.status(400).send({
-                Error: "Invalid API call."
-            })
-        }else{//when user makes excessive requests
-            res.status(400).send({
-                Error: "More than 5 trials are not supported in one minute!"
-            })
-        }
-    });
-})
+router.get("/:from-:to", teControllers.exactRate)
 
 // Retrieves all exchange rates the API can
 // uses currencylayer api
 // base currency is USD
-router.get("/$", (req, res) => {
-    request(`http://apilayer.net/api/live?access_key=${secret.currencyLayerKey.key}`, { json: true }, (err, resp, body) => {
-        if(err){
-            res.send({// if external api returns error
-                Error: "System is under maintanence."
-            });
-        }
-        else if(body["success"]){//when everything is normal sends the list of rates with their values as JSON (base currency is USD)
-            res.send(body["quotes"]);
-        }
-        else{//when user makes excessive requests
-            res.status(400).send({
-                Error: "The monthly api request limit has been reached!"
-            })
-        }
-    });
-})
+router.get("/\\$", teControllers.dollarRates)
 
  
 module.exports = router
